Added files that were deleted back. Changed what was wrong back. I hope.

+++ main/Entities/HiredIntern.java
@@ -10,7 +10,7 @@ public class  HiredIntern extends Intern {
 
     private String hiredInternName;
     private int hiredInternAge;
-    private HashMap<String, Integer> hiredInternSkills;
+    private HashMap<String, Double> hiredInternSkills;
 
     /**
      * Construct a hiredIntern, giving them the given hiredInternName,
@@ -19,7 +19,7 @@ public class  HiredIntern extends Intern {
      * @param hiredInternAge    The age of this Entities.Intern
      * @param hiredInternSkills A map of skills where the key is the name of the skill and the value is the percentage.
      */
-    public HiredIntern(String hiredInternName, int hiredInternAge, HashMap<String, Double> hiredInternSkills) {
+    public HiredIntern(String hiredInternName, int hiredInternAge, HashMap<String, Integer> hiredInternSkills) {
         super(hiredInternName, hiredInternAge, hiredInternSkills);
     }
